refactor: fix typo, reexport errors
